% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{allTheData}
\alias{allTheData}
\title{Search for all sample datasets in all currently installed packages.}
\usage{
allTheData(verbose = T)
}
\value{
A `data.frame` with columns `Package`: name of the package that
        provides that dataset, `LibPath`: path where that package is
        currently installed, `Item`: the name of the dataset, `Title`: a
        brief description of the dataset, `Class`: the class of the object
        listed in `Item` (if multiple classes, they are delimited by
        semicolons),`IsDataFrame`: whether or not the object listed in
        `Item` inherits from `data.frame`,`NumberNonNumeric`: number of
        columns that are not `numeric` (`character`, `factor`, `POSIXct`,
        etc.),`Rows`: number of rows in the `Item` if applicable,`Cols`:
        number of columns in the dataset in the `Item` if applicable.
}
\description{
A goal of this function is to be able to quickly filter through currently
available datasets and find ones that meet your needs so you're not using
the same old `mtcars` and `iris` for everything.
}
\examples{
\dontrun{ allTheData() }
}
