% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_subsample_function.R
\name{balanced_subsample}
\alias{balanced_subsample}
\title{Create random balanced subsamples}
\usage{
balanced_subsample(df, unique_vector, existing_groupings = NULL, newgroups,
  new_group_name = "new_group")
}
\arguments{
\item{df}{data.frame}

\item{unique_vector}{name of vector of unique instances}

\item{existing_groupings}{vector of column names describing groups over which to balance sampling.}

\item{newgroups}{vector of values for new groups}

\item{new_group_name}{name for variable with new groups}
}
\description{
Create random balanced subsamples
}

