% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkrun}
\alias{checkrun}
\title{Take an object name \code{obj}, check to see if it  exists in environment \code{env}
and if it does not, run \code{expression} \code{EXPR} and assign its result to \code{obj}}
\usage{
checkrun(obj, EXPR, env = as.environment(-1))
}
\arguments{
\item{obj}{A \code{character} string (required) naming the variable in env
to which the results of running \code{EXPR} will be assigned.}

\item{EXPR}{An \code{expression} to execute and assign to the object named
by \code{obj} if it doesn't already exist.}

\item{env}{An \code{environment} to check for the existence of \code{obj}
and where to create it if it doesn't exist.}
}
\value{
Does not return anything.
}
\description{
Take an object name \code{obj}, check to see if it  exists in environment \code{env}
and if it does not, run \code{expression} \code{EXPR} and assign its result to \code{obj}
}
\examples{
`checkrun('dat3',{group_by(dat1,idn_mrn) \%>\% summarise_all(first)});`
}
