% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cl_bintail}
\alias{cl_bintail}
\title{into the specified existing or new level. That level is listed last
in the resulting factor.}
\usage{
cl_bintail(xx, topn = 4, binto = "other")
}
\arguments{
\item{xx}{A \code{vector} (required).}

\item{topn}{\code{numeric} for the top how many levels to keep (optional, default =4)}

\item{binto}{\code{character} which new or existing level to dump the other values in}
}
\description{
into the specified existing or new level. That level is listed last
in the resulting factor.
}
