% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cm}
\alias{cm}
\title{takes input and returns it with a comment attribute}
\usage{
cm(xx, comment = NULL, append = T, transf = stringr::str_squish)
}
\description{
takes input and returns it with a comment attribute
}
