% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_vec_function.R
\name{cross_vec}
\alias{cross_vec}
\title{Crossed character vectors}
\usage{
cross_vec(..., FUN = function(...) paste(..., collapse = ""))
}
\arguments{
\item{...}{A set of parameters that can be coerced by as.character.}

\item{FUN}{A function to combine parameters. Defaults to \code{function(...) paste(..., collapse = "")}}
}
\value{
Character vector containing all combinations of provided parameters
}
\description{
Returns a char vector containing all combinations of provided vectors, each coerced by as.character.
}
\examples{
ovec(c("x", "y", "z"), 1:3)
# c("x1", "x2", "x3", "y1", "y2", "y3", "z1", "z2", "z3")

}
\keyword{character}
\keyword{outer,}
\keyword{vector,}

