% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filesections}
\alias{filesections}
\title{Read a file and split it according to lines containing the specified regular
expression}
\usage{
filesections(
  file,
  sectionrxp = "^#{4,}",
  targetrxp,
  subfrom = "^#*\\\\s*|\\\\s*#*$",
  subto = "",
  namefn = function(xx, subfrom, subto) {     if (grepl("^#'", xx[1]) || !grepl("^#",
    xx[1]))          ""     else {         gsub(subfrom, subto, xx[1])     } },
  ...
)
}
\arguments{
\item{file}{Name of text file to read in.}

\item{sectionrxp}{Regular expression to identify section breaks in that file}

\item{targetrxp}{Optional target expression so that only sections containing
that pattern are returned. Otherwise all sections returned.}

\item{subfrom}{Regexp indicating what to replace in order to extract the
name of each section from its first line (if that line
starts with a comment character).}

\item{subto}{Value with which to replace \code{subfrom}}

\item{namefn}{Optional function to extract names from sections. Should
be able to accept three arguments, with the first one being
a character vector representing a section of the file. The
idea is for it assign a name to the section based on its
contents.}

\item{...}{Passed to \code{readLines}}
}
\value{
List of character vectors, which may be named, but the names are not
         guaranteed legal or unique.
}
\description{
Read a file and split it according to lines containing the specified regular
expression
}
\examples{

# Note that section names are not necessarily legal variable names and not
# necessarily unique
cat('foo','bar','baz','#### section 1',123,456,789,'#### section 2'
    ,'#### section 3','bob','boo','#### section 3','bla','bip'
    ,'#### section 4','The End',sep='\\n',file='file_w_sections.txt')
filesections('file_w_sections.txt')

# keep only sections 2 and 3
filesections('file_w_sections.txt',targetrxp='^#{4} section [23]')

# keep only sections where at least one line has the string 'oo'
filesections('file_w_sections.txt',targetrxp='oo')

# keep all sections EXCEPT 'section 2'
filesections('file_w_sections.txt',targetrxp='^#{4} section [^2]')

}
