% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_filepath}
\alias{find_filepath}
\title{Find a file in some adjacent path.}
\usage{
find_filepath(
  file,
  paths = c("..", "../..", "."),
  pathexcl = c("^backup"),
  recursive = FALSE,
  lastonly = TRUE,
  normalize = TRUE
)
}
\arguments{
\item{file}{Character string for one file or file path.}

\item{paths}{Paths where to look for this file, optional character
vector.}

\item{pathexcl}{Character vector of regular expressions for excluding
paths that would otherwise match.}

\item{recursive}{Whether to look in subdirectories for this file, logical.}

\item{lastonly}{Logical. If more than one paths are found, and this is set
to \code{TRUE} (default) then only return the first path.}

\item{normalize}{Apply \code{normalizePath()} to results before returning.}
}
\value{
Character vector with path/s to file.
}
\description{
Find a file in some adjacent path.
}
\examples{
\dontrun{find_filepath('.Rprofile')}

}
