% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findheader}
\alias{findheader}
\title{Try to find a better number of initial rows to skip in a data frame like
object in order to not have columns get incorrectly coerced to character.}
\usage{
findheader(data, tryrows = 10, totalrows = 40, ...)
}
\arguments{
\item{data}{An object that inherits from \code{data.frame} or possibly
\code{matrix}}

\item{tryrows}{Number of leading rows in which to look for the header}

\item{totalrows}{The number of total rows to scan during each attempt (but
this is just for scans, the final result will always have
as many rows as possible after the one where the header is
found)}

\item{...}{Not currently used.}
}
\value{
A data.frame
}
\description{
Try to find a better number of initial rows to skip in a data frame like
object in order to not have columns get incorrectly coerced to character.
}
\examples{

foo <- unname(rbind(letters[1:5],c(NA,NA,'FOO',NA,3),names(iris)
                    ,as.matrix(iris)));
head(foo);
sapply(foo,class);

foo0 <- findheader(foo);
head(foo0);
sapply(foo,class);

}
