% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fs}
\alias{fs}
\title{Not yet ready:}
\usage{
fs(str, text = str, url = paste0("#", gsub("[^_a-z0-9]", "-",
  tolower(str))), tooltip = alist(str), class = "fl",
  template = "[\%1$s]: \%2$s \\"\%4$s\\"\\n", dct = dct0,
  col_tooltip = "colname_long", col_class = "", col_url = "",
  col_text = "", match_col = c("varname", "colname"), fs_reg = NULL,
  retfun = cat, ...)
}
\description{
Fancy Span (or any other special formatting of strings)
}
