% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getkeyval}
\alias{getkeyval}
\title{Extract key-value pairs from structured data}
\usage{
getkeyval(lines, key, joiner = "\\\\s*=\\\\s*", commentrxp = "#")
}
\arguments{
\item{lines}{Character vector (e.g. read from a file) where each
element is a line that could contain a key-value pair}

\item{key}{Character vector of key names}

\item{joiner}{Regexp for what to search for between each key and its
value}

\item{commentrxp}{Regexp for stripping out comments}
}
\value{
Named list, where the names are in \code{key}
}
\description{
Simply reads each line, finds the
}
\examples{
getkeyval(c('foo=1','bar=2','foo=DUPE','baz= 3 # comment',"bat = hello world"
            ,"# bat = old value")
          ,key=c('baz','foo','bat'))

}
