% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git}
\alias{git}
\alias{git_lsfiles}
\alias{git_other}
\alias{git_}
\alias{git_move}
\alias{git_push}
\title{Simple git wrappers}
\usage{
git_lsfiles(...)

git_other(...)

git_(...)

git_move(from, to, ...)

git_push(...)
}
\arguments{
\item{...}{Passed to the \code{git} shell command via \code{system}.}

\item{from}{Current name/path of a file tracked by git.}

\item{to}{New name/path of a file tracked by git.}
}
\value{
0 for successful execution unless \code{intern=TRUE} is part of \code{...} and if it is,
        output from the shell will be returned rather than printed to the
        screen for most commands.
}
\description{
Simple git wrappers

\code{git_lsfiles}: List only the files currently being tracked by git

\code{git_other},\code{git_}: Whatever other git functions that aren't
                              explicitly implemented yet. Just put any
                              combination of git arguments as arguments to
                              this function, leaving out \code{git} itself.

\code{git_move}: Move a file tracked by git, so git knows you did not delete it.

\code{git_push}: Push committed changes to the origin (for example, but not
necessarily, github.com)
}
\seealso{
Other git wrappers: 
\code{\link{git_add}()},
\code{\link{git_autoconf}()},
\code{\link{git_diff_filter}()},
\code{\link{git_newbranch}()},
\code{\link{git_rename}()},
\code{\link{git_status}()},
\code{\link{git_subupd}()}
}
\concept{git wrappers}
