% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_add}
\alias{git_add}
\alias{gadd}
\title{Make the specified file start getting tracked by the current git repository.}
\usage{
git_add(files, ...)

gadd(files, ...)
}
\arguments{
\item{files}{Character vector of file names.}

\item{...}{Passed to the \code{git} shell command via \code{system}.}
}
\description{
Make the specified file start getting tracked by the current git repository.
}
\seealso{
Other git wrappers: 
\code{\link{git_autoconf}()},
\code{\link{git_diff_filter}()},
\code{\link{git_newbranch}()},
\code{\link{git_rename}()},
\code{\link{git_status}()},
\code{\link{git_subupd}()},
\code{\link{git}}
}
\concept{git wrappers}
