% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_autoconf}
\alias{git_autoconf}
\title{Automatically configure your global .gitconfig with your name and email
(if not yet thus configured) so that git will allow you to commit changes}
\usage{
git_autoconf(upstream = getOption("git.upstream"), ...)
}
\description{
Automatically configure your global .gitconfig with your name and email
(if not yet thus configured) so that git will allow you to commit changes
}
