% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_checkout}
\alias{git_checkout}
\alias{gco}
\title{git checkout}
\usage{
git_checkout(which = getOption("git.workingbranch", "master"), ...)

gco(which = getOption("git.workingbranch", "master"), ...)
}
\arguments{
\item{which}{The branch you wish to check out from the current repo.}

\item{...}{Passed to the \code{git} shell command via \code{system}.}
}
\description{
git checkout
}
\seealso{
Other git wrappers: \code{\link{git_add}()},
  \code{\link{git_autoconf}()},
  \code{\link{git_diff_filter}()},
  \code{\link{git_move}()}, \code{\link{git_newbranch}()},
  \code{\link{git_status}()}, \code{\link{git_subupd}()},
  \code{\link{git}}
}
\concept{git wrappers}
