% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_diff_filter}
\alias{git_diff_filter}
\title{List the files in the repo having a particular status}
\usage{
git_diff_filter(xx)
}
\arguments{
\item{xx}{String containing one or more of A,C,D,M,R,T,U,X,B, or *}
}
\description{
Quoting from git documentation (\code{git help diff}):
\emph{Select only files that are Added (A), Copied (C), Deleted (D),
Modified (M), Renamed (R), have their type (i.e. regular file, symlink,
submodule, ...) changed (T), are Unmerged (U), are Unknown (X), or have had
their pairing Broken (B). Any combination of the filter characters (including
none) can be used. When * (All-or-none) is added to the combination, all
paths are selected if there is any file that matches other criteria in the
comparison; if there is no file that matches other criteria, nothing is
selected.}
}
