% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_exists}
\alias{git_exists}
\title{Returns \code{TRUE} if the git command is available through \code{system} and
\code{FALSE} otherwise.}
\usage{
git_exists(xx = "--version")
}
\arguments{
\item{xx}{A git command to use for testing.}
}
\value{
logical
}
\description{
Returns \code{TRUE} if the git command is available through \code{system} and
\code{FALSE} otherwise.
}
\examples{
\dontrun{git_exists()}

}
