% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_getupstream}
\alias{git_getupstream}
\title{WIP}
\usage{
git_getupstream(mergestrategy = "theirs",
  message = "Merge with upstream", fastfwd = getOption("git.fastfwd",
  F), upstrmbranch = getOption("git.upstrmbranch", "master"),
  localbranch = getOption("git.workingbranch", "master"), ...)
}
\description{
WIP
}
