% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_ignore}
\alias{git_ignore}
\title{Add a pattern to a .gitignore file}
\usage{
git_ignore(patterns, ignorepath = ".", preamble = "")
}
\arguments{
\item{patterns}{A character vector of patterns to ignore. Required.
Always appended. If you need to un-ignore something
you will have to edit .gitignore manually.}

\item{ignorepath}{Path to .gitignore (you can have multiple ones)
current directory by default.}

\item{preamble}{What to put in the line/s before a set of ignore
patterns. Empty line by default, set to NULL if you
want to not skip a line.}
}
\description{
Add a pattern to a .gitignore file
}
\examples{
git_ignore(c('*.csv','*.tsv'))
}
