% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_newbranch}
\alias{git_newbranch}
\title{Create a new branch \emph{and} check it out immediately. Optionally also
push.}
\usage{
git_newbranch(branch, pushorigin = F, ...)
}
\description{
Create a new branch \emph{and} check it out immediately. Optionally also
push.
}
