% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_other}
\alias{git_other}
\title{Whatever other git functions that aren't explicitly implemented yet. Just put
any combination of git arguments as arguments to this function, leaving out
\code{git} itself.}
\usage{
git_other(...)
}
\description{
Whatever other git functions that aren't explicitly implemented yet. Just put
any combination of git arguments as arguments to this function, leaving out
\code{git} itself.
}
