% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_push}
\alias{git_push}
\title{Push committed changes to the origin (for example, but not necessarily,
github.com)}
\usage{
git_push(...)
}
\description{
Push committed changes to the origin (for example, but not necessarily,
github.com)
}
