% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_rename}
\alias{git_rename}
\title{Rename a file tracked by git, so git knows you did not delete it.}
\usage{
git_rename(from, to, ...)
}
\arguments{
\item{from}{Current name/path of a file tracked by git.}

\item{to}{New name/path of a file tracked by git.}

\item{...}{Passed to the \code{git} shell command via \code{system}.}
}
\description{
Rename a file tracked by git, so git knows you did not delete it.
}
\seealso{
Other git wrappers: \code{\link{git_add}()},
  \code{\link{git_autoconf}()},
  \code{\link{git_checkout}()},
  \code{\link{git_diff_filter}()},
  \code{\link{git_move}()}, \code{\link{git_newbranch}()},
  \code{\link{git_status}()}, \code{\link{git_subupd}()},
  \code{\link{git}}
}
\concept{git wrappers}
