% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_ssh}
\alias{git_ssh}
\title{Switch between ssh authentication and ssl authentication for a git repo.}
\usage{
git_ssh(tossh = TRUE, sshstr = "git@github.com:",
  sslstr = "https://github.com/")
}
\arguments{
\item{tossh}{If `TRUE`, will attempt to convert the remote.origin.url from
https to ssh. Default: `TRUE`}

\item{sshstr}{A string to use as the prefix for ssh connection. Optional,
defaults to the values used by github.com}

\item{sslstr}{A string to use as the prefix for the ssl connection. Optional,
defaults to the values used by github.com.}
}
\value{
Invisibly returns `0` or an error code.
}
\description{
A use-case for this is some environments that by default initialize projects
as ssl/https (e.g. RStudio Cloud) but some users may prefer ssh
authentication. This easily converts between the two settings without having
to remember the whole git command. Will silently return errors from shell but
will not throw an error.
}
\examples{
\dontrun{
# Convert from https://github.com/... to git@github.com:...
git_ssh()

# Convert from git@github.com:... to https://github.com/...
git_ssh(FALSE)

}
}
