% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_status}
\alias{git_status}
\title{Nicely formatted and concise status of current git repo.}
\usage{
git_status(print = T, diff_filters = list(Added = "A", Copied = "C",
  Deleted = "D", Modified = "M", Renamed = "R", ChangedType = "T", Unmerged
  = "U", Unknown = "X", Broken = "B"), ...)
}
\description{
Nicely formatted and concise status of current git repo.
}
