% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_subupd}
\alias{git_subupd}
\title{Delete and re-download git submodules if any exist.}
\usage{
git_subupd(stopfile = ".developer")
}
\arguments{
\item{stopfile}{The name of a file which, if exists, will cause this function
to exit without doing anything. Will silently return errors 
from shell but will not throw an error.}
}
\value{
If successful, \code{0}, otherwise an error code.
}
\description{
Delete and re-download git submodules if any exist.
}
\examples{

\dontrun{ git_subupd() }
}
