% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grepor}
\alias{grepor}
\title{Take a data.frame or character vector and a vector of grep targets and return
the values that match (for data.frame, column names that match). If no
patterns given just returns the names}
\usage{
grepor(xx, patterns = ".")
}
\arguments{
\item{xx}{A \code{data.frame} or character vector (required)}

\item{patterns}{A character vector of regexp targets to be OR-ed}
}
\description{
Take a data.frame or character vector and a vector of grep targets and return
the values that match (for data.frame, column names that match). If no
patterns given just returns the names
}
