% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_deps}
\alias{load_deps}
\title{Load or render script dependencies if not already cached.}
\usage{
load_deps(
  deps,
  scriptdir = getwd(),
  cachedir = scriptdir,
  fallbackdir = "scripts",
  envir = parent.frame(),
  loadfn = if (exists("tload")) tload else load,
  rendfn = "rmarkdown::render",
  ...
)
}
\arguments{
\item{deps}{Character vector of script names.}

\item{scriptdir}{Where to look for each script.}

\item{cachedir}{Where to save the \code{SCRIPTNAME.R.rdata} output from
each script and where to first look for cached results if
they exist.}

\item{fallbackdir}{Where to look for each script that is not found in
\code{scriptdir}.}

\item{envir}{Environment in which to evaluate scripts (recommend
leaving unaltered).}

\item{loadfn}{Function for loading \code{SCRIPTNAME.R.rdata} cached
results}

\item{rendfn}{String with name of function to run on each of the files.
If it's not from the \code{base} library it should be
fully qualified (i.e. like the default value). The two
anticipated values for this argument are
\code{'rmarkdown::render'} for generating reports together
with cached output and \code{'source'} for generating just
the output. Other functions might also do useful things,
but no guarantees.}

\item{...}{Arguments to pass to the function specified in
\code{rendfn}}
}
\value{
Character vector of objects created and saved by the scripts that
        have been loaded into the working environment.
}
\description{
Load or render script dependencies if not already cached.
}
