% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapnames}
\alias{mapnames}
\title{Usage: `xx<-mapnames(xx,lookup)` where lookup is a named character vector
the names of the elements in the character vector are what you are renaming
things TO and the values are what needs to be matched, i.e. what renaming things
FROM. If you set namesonly=T then it just returns the names, not the original
object.}
\usage{
mapnames(xx, lookup, namesonly = F, ...)
}
\description{
Usage: `xx<-mapnames(xx,lookup)` where lookup is a named character vector
the names of the elements in the character vector are what you are renaming
things TO and the values are what needs to be matched, i.e. what renaming things
FROM. If you set namesonly=T then it just returns the names, not the original
object.
}
