% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_last_function.R
\name{match_last}
\alias{match_last}
\title{Match on last element}
\usage{
match_last(x, table, nomatch = NA_integer_, incomparables = NULL)
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched. \code{\link{Long vectors}} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against. \code{\link{Long vectors}} are not supported.}

\item{nomatch}{the value to be returned in the case when no match is found. Note that it is coerced to \code{integer}.}

\item{incomparables}{a vector of values that cannot be matched. Any value in \code{x} matching a value in this vector is assigned the \code{nomatch} value. For historical reasons, \code{FALSE} is equivalent to \code{NULL}.}
}
\value{
Performs a \code{\link{match}}, but on the last element matching, rather than the first.
}
\description{
\code{\link{match}} returns a vector of the positions of (first) matches of its first argument in its second.
\code{\link{match_last}} returns a vector of the positions of (last) matches of its first argument in its second.
}

