% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovec_function.R
\name{ovec}
\alias{ovec}
\title{Function for generating a character vector from crossing vectors, separated by sep}
\usage{
ovec(..., FUN = function(a, b) paste0(a, b))
}
\arguments{
\item{...}{A set of parameters that can be interpreted as character vectors.}

\item{FUN}{A function to combine parameters. Defaults to function(a, b) paste0(a, b)}
}
\description{
Function for generating a character vector from crossing vectors, separated by sep
}
\examples{
ovec(c("x", "y", "z"), 5:10)
}
\keyword{character}
\keyword{outer,}
\keyword{vector,}

