% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages_function.R
\name{packages}
\alias{packages}
\title{Shortcut to install packages}
\usage{
packages(x)
}
\arguments{
\item{x}{name of package to be installed/loaded}
}
\value{
Returns nothing.
}
\description{
Loads package x if it exists, otherwise tries to download x.
}
\keyword{install.packages,}
\keyword{library}

