% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_wd.R
\name{set_wd}
\alias{set_wd}
\title{prints template for fetching the directory associated with a given file}
\usage{
set_wd()
}
\description{
prints template for fetching the directory associated with a given file
}
