% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfig_function.R
\name{sigfig}
\alias{sigfig}
\title{Only selected number of digits}
\usage{
sigfig(x, digits = 3)
}
\arguments{
\item{x}{numeric vector.}

\item{digits}{number of digits to return for each number.}
}
\value{
returns a vector of same length as x, with number of digits = \code{digits}
}
\description{
Only selected number of digits
}

