% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{submulti}
\alias{submulti}
\title{Take a character vector and perform multiple search-replace
operations on it.}
\usage{
submulti(
  xx,
  searchrep,
  method = c("partial", "full", "exact", "starts", "ends", "startsends")
)
}
\arguments{
\item{xx}{A \code{vector} of type \code{character} (required)}

\item{searchrep}{A \code{matrix} with two columns of type \code{character} (required). The left column is the pattern and the right, the replacement.}

\item{method}{One of 'partial','full', or 'exact'. Controls whether to replace only the matching regexp, replace the entire value that contains a matching regexp, or replace the entire value if it's an exact match.}
}
\description{
Take a character vector and perform multiple search-replace
operations on it.
}
