% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{systemRootDir}
\alias{systemRootDir}
\title{Take a set of objects coercible to matrices and perform sprintf on them while
preserving their dimensions (obtained from the first argument of ...)}
\usage{
systemRootDir()
}
\description{
Take a set of objects coercible to matrices and perform sprintf on them while
preserving their dimensions (obtained from the first argument of ...)
}
