% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tblinfo}
\alias{tblinfo}
\title{Create an automated data dictionary}
\usage{
tblinfo(dat, custom_stats = alist(), info_cols = alist(c_empty =
  frc_missing == 1, c_uninformative = n_nonmissing < 2, c_ordinal =
  uniquevals < 10 & isnum, c_tm = uniquevals == 1 & n_missing > 0, c_tf =
  uniquevals == 2, c_numeric = isnum & !c_ordinal, c_factor = uniquevals <
  20 & !isnum, c_complex = !(c_ordinal | c_tm | c_tf | c_numeric |
  c_factor)), ...)
}
\arguments{
\item{dat}{An object that inherits from `data.frame`}

\item{custom_stats}{An `alist` of statistics to calculate on each
column of `dat` in addition to the defaults
in `info_cols` (below). Optional.}

\item{info_cols}{Another `alist`, this one has default values but
can be overridden on an all-or-none basis.}

\item{...}{Eats any extra arguments, to keep them from
causing trouble.}
}
\value{
A data-frame having one row for each column in `dat`
}
\description{
Create an automated data dictionary
}
\examples{

tblinfo(datasets::iris)

}
