% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/this_fp.R
\name{this_fp}
\alias{this_fp}
\title{prints template for fetching the directory associated with a given file}
\usage{
this_fp()
}
\description{
prints template for fetching the directory associated with a given file
}
