% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidbits-package.r
\docType{package}
\name{tidbits}
\alias{tidbits}
\title{tidbits.}
\description{
Various functions to streamline tasks that are repetitive and/or require
funky code tricks. A byproduct of UTHealth's TSCI 5050 Introduction to Data
Science course.
}
