% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{truthy}
\alias{truthy}
\title{Example of using R methods dispatch}
\usage{
truthy(xx, ...)
}
\description{
The actual usage is: `truthy(foo)` and `truthy()` itself figures
out which method to actually dispatch.
}
