% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulstrsplit_function.R
\name{ulstrsplit}
\alias{ulstrsplit}
\title{Unlist strsplit}
\usage{
ulstrsplit(...)
}
\arguments{
\item{...}{Parameters to be forwarded to \code{\link{strsplit}}}
}
\value{
Unlist strsplit
}
\description{
Unlist strsplit
}
\examples{
x <- c(as = "asfef", qu = "qwerty", "yuiop[", "b", "stuff.blah.yech")
#split x on the letter e
ulstrsplit(x, "e")
}

