% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vartype}
\alias{vartype}
\title{Returns a vector of column names that contain data elements of a particular type
as specified by the user: "integer","POSIXct" "POSIXt", "numeric", "character", 
"factor" and "logical".}
\usage{
vartype(dat, ctype)
}
\description{
Returns a vector of column names that contain data elements of a particular type
as specified by the user: "integer","POSIXct" "POSIXt", "numeric", "character", 
"factor" and "logical".
}
