% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vec2m}
\alias{vec2m}
\title{Stack a vector to form a matrix with repeating rows, with optional
column names and transformation}
\usage{
vec2m(vv, nr = 1, trans = identity)
}
\arguments{
\item{vv}{A vector which will become the row of a matrix}

\item{nr}{Number of (identical) rows this matrix will contain}

\item{trans}{An optional function that can take a matrix as its
sole argument. Useful functions include `as.data.frame()`
`as_tibble()` and `as.table()`}
}
\value{
A matrix, unless the function specified in the `trans` argument
        causes the output to be something else.
}
\description{
Stack a vector to form a matrix with repeating rows, with optional
column names and transformation
}
\examples{
vec2m(1:10,5);
vec2m(1:10,5,tr=data.frame);
vec2m(setNames(1:12,month.name),3);
}
