% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{with_attrs}
\alias{with_attrs}
\title{Append or replace attributes of any object in a pipeline-frindly way.}
\usage{
with_attrs(xx, rep.attrs = list(), app.attrs = list())
}
\arguments{
\item{xx}{Object whose attributes to modify and then return the object}

\item{rep.attrs}{Named list of attributes to create or replace}

\item{app.attrs}{Named list of attributes to create or append to}
}
\value{
The object `xx` with modified attributes
}
\description{
Append or replace attributes of any object in a pipeline-frindly way.
}
\examples{
# Change an object's attribute
with_attr(iris,list(class='list'))

# Create a new attribute for an object
foo <- with_attr(LETTERS,list(comment='Hello world'))
comment(foo)
foo <- with_attr(foo,rep='One more comment.')
comment(foo)

}
