% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{with_cm}
\alias{with_cm}
\alias{cm}
\title{takes input and returns it with a comment attribute}
\usage{
with_cm(xx, comment = NULL, append = TRUE, transf = stringr::str_squish)

cm(xx, comment = NULL, append = TRUE, transf = stringr::str_squish)
}
\arguments{
\item{xx}{Any R object}

\item{comment}{Comment to add to the object.}

\item{append}{If \code{TRUE} (default) appends comment to those that
are already attached to the object, if any.}

\item{transf}{Function that takes one character vector argument and returns
a character vector. By default it is
\code{\link[stringr]{str_squish}} which removes extra whitespace.
To bypass entirely set this argument to \code{NULL}.}
}
\value{
The same object passed as the first argument but with the added
        comment.
}
\description{
takes input and returns it with a comment attribute
}
\examples{
data(iris)
comment(iris)
foo <- cm(iris,comment=' What  pretty ')
comment(foo)

foo <- cm(foo,comment='flowers!',transf=toupper)
comment(foo)

# This replaces the earlier comments due to append being FALSE now.
foo <- cm(foo,'Hello world.',append=FALSE)
comment(foo)

}
