% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{autoread}
\alias{autoread}
\title{Autoguessing function for reading most common data formats}
\usage{
autoread(file, na = c("", ".", "(null)", "NULL", "NA"),
  fixnames = function(xx) {     setNames(xx,
  tolower(make.names(names(xx)))) }, file_args = list(), ...)
}
\description{
Autoguessing function for reading most common data formats
}
