% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findheader}
\alias{findheader}
\title{Extends trailR package with integrated universal (almost) file reader}
\usage{
findheader(data, tryrows = 10, totalrows = 40, ...)
}
\arguments{
\item{data}{An object that inherits from \code{data.frame} or possibly
\code{matrix}}

\item{tryrows}{Number of leading rows in which to look for the header}

\item{totalrows}{The number of total rows to scan during each attempt (but
this is just for scans, the final result will always have
as many rows as possible after the one where the header is
found)}

\item{...}{Not currently used.}

\item{file}{Any of the common delimited file formats}
}
\value{
Whatever object the underlying read function passed to `tread()`
        will output, usually a class that inherits from `data.frame`

export
Try to find a better number of initial rows to skip in a data frame like
object in order to not have columns get incorrectly coerced to character.

A data.frame
}
\description{
Extends trailR package with integrated universal (almost) file reader
}
\examples{

foo <- unname(rbind(letters[1:5],c(NA,NA,'FOO',NA,3),names(iris)
                    ,as.matrix(iris)));
head(foo);
sapply(foo,class);

foo0 <- findheader(foo);
head(foo0);
sapply(foo,class);

}
