% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_add}
\alias{git_add}
\title{Make the specified file start getting tracked by the current git repository.}
\usage{
git_add(files, ...)
}
\description{
Make the specified file start getting tracked by the current git repository.
}
