% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_status}
\alias{git_status}
\alias{gst}
\title{Nicely formatted and concise status of current git repo.}
\usage{
git_status(
  print = TRUE,
  diff_filters = list(Added = "A", Copied = "C", Deleted = "D", Modified = "M", Renamed
    = "R", ChangedType = "T", Unmerged = "U", Unknown = "X", Broken = "B"),
  ...
)

gst(
  print = TRUE,
  diff_filters = list(Added = "A", Copied = "C", Deleted = "D", Modified = "M", Renamed
    = "R", ChangedType = "T", Unmerged = "U", Unknown = "X", Broken = "B"),
  ...
)
}
\arguments{
\item{print}{If \code{TRUE} (default) prints formatted results to
console.}

\item{diff_filters}{Which statuses to display (all of them, by default)}

\item{...}{Passed to the \code{git} shell command via \code{system}.}
}
\description{
Nicely formatted and concise status of current git repo.
}
\seealso{
Other git wrappers: 
\code{\link{git_add}()},
\code{\link{git_autoconf}()},
\code{\link{git_diff_filter}()},
\code{\link{git_newbranch}()},
\code{\link{git_rename}()},
\code{\link{git_subupd}()},
\code{\link{git}}
}
\concept{git wrappers}
