% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_subupd}
\alias{git_subupd}
\title{Delete and re-download git submodules if any exist.}
\usage{
git_subupd(stopfile = ".developer")
}
\arguments{
\item{stopfile}{The name of a file which, if exists, will cause this function
to exit without doing anything. Will silently return errors
from shell but will not throw an error.}
}
\value{
0 for successful execution unless \code{intern=TRUE} is part of \code{...} and if it is,
        output from the shell will be returned rather than printed to the
        screen for most commands.
}
\description{
Delete and re-download git submodules if any exist.
}
\examples{

\dontrun{ git_subupd() }

}
\seealso{
Other git wrappers: \code{\link{git_add}()},
  \code{\link{git_autoconf}()},
  \code{\link{git_checkout}()},
  \code{\link{git_diff_filter}()},
  \code{\link{git_move}()}, \code{\link{git_newbranch}()},
  \code{\link{git_rename}()}, \code{\link{git_status}()},
  \code{\link{git}}
}
\concept{git wrappers}
