% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{instrequire}
\alias{instrequire}
\title{This function takes a list of package names, loads them if they are
available, otherwise attempts to install each one and then again
attempts to load it.}
\usage{
instrequire(pkgs, quietly = TRUE, dependencies = TRUE,
  repos = getOption("repos", "https://cran.rstudio.com/"), ...)
}
\description{
This function takes a list of package names, loads them if they are
available, otherwise attempts to install each one and then again
attempts to load it.
}
