% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{instrequire}
\alias{instrequire}
\title{try to install and load each of a list of packages}
\usage{
instrequire(
  pkgs,
  quietly = TRUE,
  dependencies = TRUE,
  type = if (get_os() \%in\% c("osx", "windows")) {     "binary" } else "source",
  checksource = "no",
  compile = "never",
  repos = getOption("repos", "https://cran.rstudio.com/"),
  ...
)
}
\arguments{
\item{pkgs}{character vector of package names that are needed, e.g.
by a script that uses this command.}

\item{quietly}{passed to \code{\link[base]{require}}}

\item{dependencies}{whether to also install dependencies, passed to
\code{\link[utils]{install.packages}}}

\item{type}{passed to \code{link[utils]{install.packages}}. By
default it is \code{'binary'} for all MacOS and Windows
and \code{'source'} otherwise}

\item{checksource}{temporarily sets
\code{options('install.packages.check.source')}
to this value}

\item{compile}{temporarily sets
\code{options('install.packages.compile.from.source')}
to this value}

\item{repos}{passed to \code{\link[utils]{install.packages}} to
specify which repositories to use (optional).}

\item{...}{passed to \code{\link[utils]{install.packages}} to
specify any additional options needed (e.g. local urls
and such).}
}
\description{
This function takes a list of package names, loads them if they are
available, otherwise attempts to install each one and then again
attempts to load it. Will produce an error if any packages in \code{pkgs}
are not possible to install and load.
}
\details{
The goal is to provide a robust and automation-friendly alternative to
putting a bunch of \code{\link[base]{require}} or \code{\link[base]{library}}
statements at the start of each script and have them keep erroring until each
dependency is manually installed.
}
\examples{
instrequire(c('readr','readxl','tibble'))
}
