% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovec_function.R
\name{ovec}
\alias{ovec}
\title{Crossed character vectors}
\usage{
ovec(..., FUN = function(a, b) paste0(a, b))
}
\arguments{
\item{...}{A set of parameters that can be coerced by as.character.}

\item{FUN}{A function to combine parameters. Defaults to \code{function(a, b) paste0(a, b)}}
}
\value{
Character vector containing all combinations of provided parameters
}
\description{
Crossed character vectors
}
\examples{
ovec(c("x", "y", "z"), 1:3)
# c("x1", "x2", "x3", "y1", "y2", "y3", "z1", "z2", "z3")
}
\keyword{character}
\keyword{outer,}
\keyword{vector,}

