% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{t_autoread}
\alias{t_autoread}
\title{Extends trailR package with integrated universal (almost) file reader}
\usage{
t_autoread(file, ...)
}
\arguments{
\item{file}{Any of the common delimited file formats}
}
\description{
Extends trailR package with integrated universal (almost) file reader
}
