% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tblinfo}
\alias{tblinfo}
\title{Create an automated data dictionary}
\usage{
tblinfo(
  dat,
  custom_stats = alist(),
  info_cols = alist(c_empty = frc_missing == 1, c_uninformative = n_nonmissing < 2,
    c_ordinal = uniquevals < min(10, nn) & uniquevals > 0 & isnum, c_tm = uniquevals == 1
    & n_missing > 0, c_tf = uniquevals == 2, c_numeric = isnum & !c_ordinal, c_factor =
    uniquevals < min(10, nn) & !isnum, c_safe = frc_missing < 0.2 & frc_mostcommon < 0.7
    & !c_uninformative & !(!c_numeric & uniquevals > 15), c_safetf = c_safe & c_tf,
    c_safenumeric = c_safe & c_numeric, c_complex = !(c_ordinal | c_tm | c_tf | c_numeric
    | c_factor)),
  ...
)
}
\arguments{
\item{dat}{An object that inherits from \code{data.frame}}

\item{custom_stats}{An \code{alist} of statistics to calculate on each
column of \code{dat} in addition to the defaults
in \code{info_cols} (below). Optional.}

\item{info_cols}{Another \code{alist}, this one has default values but
can be overridden on an all-or-none basis.}

\item{...}{Eats any extra arguments, to keep them from
causing trouble.}
}
\value{
A data-frame having one row for each column in \code{dat}
}
\description{
The unevaluated expressions in \code{custom_stats} and \code{info_cols} use
variables that exist during execution and are documented in
\code{\link{colinfo}}
}
\examples{

tblinfo(datasets::iris)

}
\seealso{
v

colinfo
}
